refactor: Send customization options by constructor without inheritance
